Matrix multiplication added and indexing setup
Whenever we took input from the user, we had to decrease it in every place we want the matrix to be. So instead of changing it everywhere I just decreased the value of mat just as we take input from the user. That's why in while condition we had to use -2 instead of -1 because value is decreased by 1.